return logs found

+++ migrate_helper_scripts/archive_logs.py
@@ -1,10 +1,9 @@
 #!/home/users/jeffderb/python3/bin/python3
 
-import migrate_helper_scripts.list_logs as list_logs
 import os
 import gzip
 import shutil
-import pprint
+import migrate_helper_scripts.list_logs as list_logs
 
 
 LOG_DIRECTORY = "/var/migration/"
@@ -20,20 +19,23 @@ def get_year_month(log_name):
 def archive(command="archive", volumes=False):
     logs = list_logs.get_logs(command, volumes)
     log_total = archived = removed = 0
-    for log in logs:
-        year_month = get_year_month(log)
-        os.makedirs(LOG_DIRECTORY + ARCHIVE_DIR + year_month, exist_ok=True)
-        log_file_path = LOG_DIRECTORY + log
-        if os.path.exists(log_file_path):
-            log_total += 1
-            gz_file_path = LOG_DIRECTORY + ARCHIVE_DIR + year_month + '/' + log + '.gz'
-            with open(log_file_path, 'rb') as f_in:
-                with gzip.open(gz_file_path, 'wb') as f_out:
-                    shutil.copyfileobj(f_in, f_out)
-                    archived += 1
-            if os.path.exists(gz_file_path):
-                os.remove(log_file_path)
-                removed += 1
+    if len(logs) > 0:
+        for log in logs:
+            year_month = get_year_month(log)
+            os.makedirs(LOG_DIRECTORY + ARCHIVE_DIR + year_month, exist_ok=True)
+            log_file_path = LOG_DIRECTORY + log
+            if os.path.exists(log_file_path):
+                log_total += 1
+                gz_file_path = LOG_DIRECTORY + ARCHIVE_DIR + year_month + '/' + log + '.gz'
+                with open(log_file_path, 'rb') as f_in:
+                    with gzip.open(gz_file_path, 'wb') as f_out:
+                        shutil.copyfileobj(f_in, f_out)
+                        archived += 1
+                if os.path.exists(gz_file_path):
+                    os.remove(log_file_path)
+                    removed += 1
+    else:
+        return {'logs found': logs}
 
     return {'log count': log_total, 'archived': archived, 'removed': removed}
